pyproject.toml
[project]
name = "NJK"
version = "0.1.0"
dependencies = [
    "imagehash>=4.3.2",
    "openai>=2.12.0",
    "peewee>=3.18.3",
    "peewee-migrate>=1.14.3",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "websockets>=15.0.1",
]
